LayoutTests/http/tests/inspector/network/network-redirect.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script>
function loadIframe()
{
    var iframe = document.createElement("iframe");
    iframe.src = "resources/redirect.php?status=302";
    iframe.onload = iframeLoaded;
    document.body.appendChild(iframe);
}

function iframeLoaded()
{
    console.log("Done.");
}

function test()
{
    InspectorTest.addSniffer(WebInspector.ConsoleView.prototype, "addMessage", step2);
    InspectorTest.evaluateInPage("loadIframe()");

    function step2()
    {
        // inspector-test.js appears in network panel occasionally in Safari on
        // Mac, so checking two last resources.

        InspectorTest.addResult("Model data:");

        var resources = WebInspector.panels.network.resources;
        for (var i = resources.length - 2; i < resources.length; ++i)
            printResource(resources[i]);

        InspectorTest.addResult("Grid data:");

        var dataGridNodes = WebInspector.panels.network._dataGrid.children;
        WebInspector.panels.network.refresh();
        for (var i = dataGridNodes.length - 2; i < dataGridNodes.length; ++i)
            printResource(dataGridNodes[i]._resource);
        InspectorTest.completeTest();
    }

    function printResource(resource)
    {
        InspectorTest.addResult(resource.requestMethod + ": " + resource.url + " -> " + resource.statusCode);
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests network panel redirect.</p>
</body>
</html>
